% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detrend.R
\name{Detrend}
\alias{Detrend}
\title{Detrend
Simple function fits a straight line to a vector of values, and uses the
regression coefficients to subtract off the linear trend from the values.}
\usage{
Detrend(pt)
}
\arguments{
\item{pt}{A vector of numeric values to be detrended}
}
\value{
A vector of detrended values
}
\description{
Detrend
Simple function fits a straight line to a vector of values, and uses the
regression coefficients to subtract off the linear trend from the values.
}
