% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetrendHeight.R
\name{DetrendHeight}
\alias{DetrendHeight}
\title{Remove trend from a time series}
\usage{
DetrendHeight(pt)
}
\arguments{
\item{pt}{A vector of numeric values to be detrended}
}
\value{
A vector of detrended values
}
\description{
Fits a straight line to a vector of values using lm(), and uses the
regression coefficients to subtract off the linear trend from the values.
Returns a series of residuals around the linear trend.
Typically this is used to remove a tidal trend from a ocean 
surface height time series before attempting to calculate statistics for 
waves, which are variations above and below the mean surface height.
}

