% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinOWHLfiles.R
\name{joinOWHLfiles}
\alias{joinOWHLfiles}
\title{Concatenate multiple OWHL raw csv data files}
\usage{
joinOWHLfiles(filedir, timezone = "UTC")
}
\arguments{
\item{timezone}{Specifies the time zone the raw data timestamps represent. UTC is 
the preferred time zone for simplicity.}

\item{dir}{A path to a directory containing OWHL csv files.}
}
\value{
A data frame with the individual csv files concatenate together
}
\description{
Concatenate multiple OWHL raw csv data files
}
\details{
Concatenate multiple OWHL raw csv data files into a single data frame and 
interpolate any missing 1-second intervals. The resulting data frame can 
be used to generate wave statistics from other functions in the package.

The input csv files can have mission info in the first row, and then the 
following columns: POSIXt, DateTime, frac.seconds, Pressure.mbar, TempC
}
