% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWaveSpectrum.R
\name{plotWaveSpectrum}
\alias{plotWaveSpectrum}
\title{Plot a basic smoothed spectrum}
\usage{
plotWaveSpectrum(PT, Fs)
}
\arguments{
\item{PT}{A vector of surface heights}

\item{Fs}{Frequency of sampled surface heights, units of Hz}
}
\description{
Plot a basic smoothed spectrum
}

