% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveNumL.R
\name{WaveNumL}
\alias{WaveNumL}
\title{WaveNumL}
\usage{
WaveNumL(f, h)
}
\arguments{
\item{f}{A numeric vector of wave frequencies}

\item{h}{A numeric vector of water depths (usually in units of meters)}
}
\value{
The wave number.
}
\description{
A function to calculate wavenumber.
}
\references{
Original MATLAB function by Urs Neumeier 
http://neumeier.perso.ch/matlab/waves.html
}
\author{
George Voulgaris, SUDO, 1992
}
