% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavesp.R
\name{wavesp}
\alias{wavesp}
\title{Calculate wave parameters using spectral analysis}
\usage{
wavesp(PT, Fs)
}
\arguments{
\item{PT}{A vector of surface heights that constitute a time series of 
observations. Units = meters}

\item{Fs}{Sampling frequency of the surface heights data. Units = Hz, i.e.
samples per second.}
}
\value{
Data frame of wave parameters based on spectral methods
}
\description{
Calculate wave parameters using spectral analysis
}

