% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveStatsZC.R
\name{almostZero}
\alias{almostZero}
\title{Test whether vector elements are effectively zero}
\usage{
almostZero(x, y = 0, tolerance = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{Numeric vector of values to compared against 0}

\item{y}{Value to be compared against. Default is 0.}

\item{tolerance}{The maximum difference between x and y necessary to call a 
value non-zero}
}
\value{
TRUE or FALSE for each element of the vector x.
}
\description{
Test whether elements in vector x are effectively zero,
within the square root of machine tolerance for a floating point number.
}
\details{
Returns TRUE for all vector elements that are within rounding error of zero.
}

