% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWaveSpectrum.R
\name{plotWaveSpectrum}
\alias{plotWaveSpectrum}
\title{Plot a basic smoothed spectrum}
\usage{
plotWaveSpectrum(freqspec, Fs)
}
\arguments{
\item{freqspec}{A data frame containing a column of frequencies and a
column of spectral power values}

\item{Fs}{Frequency of sampled surface heights, units of Hz}
}
\description{
Plot a basic smoothed spectrum
}

