% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prCorr.R
\name{prCorr}
\alias{prCorr}
\title{Correct for depth attenuation of a water surface height 
pressure signal.}
\usage{
prCorr(pt, Fs, zpt, M = 512, CorrLim = c(0.05, 0.33), plot = FALSE)
}
\arguments{
\item{pt}{A vector of water surface heights (units of meters usually).}

\item{Fs}{Sampling frequency (units of Hz). Normally 4Hz for an OWHL logger.}

\item{zpt}{Height of the pressure sensor above the seabed (units of meters).}

\item{M}{Length of time series segments that will be used in the detrending
and attenuation correction operations. 512 samples is the default, should be
an even number.}

\item{CorrLim}{[min max] frequency for attenuation correction (Hz, optional, 
default [0.05 0.33]).}
}
\value{
A vector of the depth-corrected surface heights (units of meters 
usually). Any original trend in the input data (such as tide change) is 
present in the output data. The returned surface height fluctuations will 
typically be more extreme than the raw input surface heights.
}
\description{
Bottom-mounted pressure transducers suffer from pressure signal 
attenuation when attempting to estimate surface wave heights. This 
function corrected water surface height time series based on the
depth of the water column and height of the sensor above the bottom.
}
\examples{
data(wavedata)
corrected = prCorr(wavedata$SurfaceHeightRaw.m, Fs = 4, zpt = 0.1)
# Plot the results
plot(x = wavedata$DateTime, y = corrected, type = 'l', 
 ylab='Surface Height, m', xlab = 'Time')
lines(x = wavedata$DateTime, y = wavedata$SurfaceHeightRaw.m, col = 'red')
legend('topleft',legend=c('Corrected','Raw'),col=c('black','red'),lwd = 2)
}
\references{
Original MATLAB function by Urs Neumeier 
http://neumeier.perso.ch/matlab/waves.html Modified from the Pcorr3.m 
function written by T. Mason, SOC, January 1997
}

