% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prCorr.R
\name{waveNumL}
\alias{waveNumL}
\title{A function to calculate wave number.}
\usage{
waveNumL(f, h)
}
\arguments{
\item{f}{A numeric vector of wave frequencies}

\item{h}{A numeric vector of water depths (usually in units of meters)}
}
\value{
The wave number.
}
\description{
A function to calculate wave number.
}
\references{
Modified from MATLAB function by Urs Neumeier:  
http://neumeier.perso.ch/matlab/waves.html
Based on polynomial approximation from Hunt, J. N. 1979. “Direct Solution of 
Wave Dispersion Equation,” ASCE Journal of the Waterway, Port, Coastal and Ocean Division.,
Vol 105, pp 457-459 suggested by George Voulgaris, University of South Carolina
}
