% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prCorr.R
\name{waveNumL}
\alias{waveNumL}
\title{A function to calculate wave number.}
\usage{
waveNumL(f, h)
}
\arguments{
\item{f}{A numeric vector of wave frequencies}

\item{h}{A numeric vector of water depths (usually in units of meters)}
}
\value{
The wave number.
}
\description{
A function to calculate wave number.
}
\references{
Modified from MATLAB function by Urs Neumeier:  
http://neumeier.perso.ch/matlab/waves.html
}
\author{
George Voulgaris, SUDO, 1992
}
