% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveStatsSP.R
\name{waveStatsSP}
\alias{waveStatsSP}
\title{Calculate ocean wave parameters using spectral analysis}
\usage{
waveStatsSP(PT, Fs)
}
\arguments{
\item{PT}{A vector of surface heights that constitute a time series of 
observations. Units = meters}

\item{Fs}{Sampling frequency of the surface heights data. Units = Hz, i.e.
samples per second.}
}
\value{
Data frame of wave parameters based on spectral methods
}
\description{
Calculate ocean wave parameters using spectral analysis
}
\details{
Carries out spectral analysis of ocean wave height time series to estimate
common wave height statistics, including peak period, average period, 
and significant wave height.
}

