% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveStatsZC.R
\name{waveStatsZC}
\alias{waveStatsZC}
\title{Calculate wave statistics using zero-cross method}
\usage{
waveStatsZC(data, Fs, threshold = NULL, plot = FALSE)
}
\arguments{
\item{data}{A numeric vector of water surface height data. The data do not
need to be detrended prior to use.}

\item{Fs}{Sampling frequency of the data, in Hz.}

\item{threshold}{The minimum height necessary for a zero-crossing event to 
be considered a wave.}

\item{plot}{Set to TRUE if summary histograms of wave heights and wave
periods are desired.}
}
\value{
A list object containing summary statistic values.
}
\description{
A function to calculate ocean wave summary statistics, including significant
wave height and period, based on an upward zero-crossing algorithm originally
provided by Urs Neumeier, v1.06.
}
\examples{
data(wavedata)
waveStatsZC(data = wavedata$SurfaceHeight.m, Fs = 4)
}
\references{
Original MATLAB function by Urs Neumeier:  
http://neumeier.perso.ch/matlab/waves.html
}

